fix: code smell

+++ src/routes/(ui)/(plain)/logout/+server.ts
@@ -6,9 +6,7 @@ export const GET: RequestHandler = async ({ locals }) => {
 	const sessionService = locals.container.resolve('sessionService');
 	if (locals.authentication.type === 'SESSION') {
 		const sessionId = locals.authentication.sessionId;
-		if (sessionId) {
-			await sessionService.deleteSession(sessionId);
-		}
+		if (sessionId) await sessionService.deleteSession(sessionId);
 	}
 	throw redirect(302, '/');
 };
